puts the application routes under an `(application)` grouping to separate layout from dev route (#979)
* generalizes the application layout as its own component

* separates application layout via an application layout group, preserving other routes to do their thing

* update comment

* moves around new layout components into more coherent location

* fixes build error
