+++ web-local/global.d.ts
@@ -1,1 +1,5 @@
 /// <reference types="@sveltejs/kit" />
+interface ImportMetaEnv {
+  RILL_VERSION: string;
+  RILL_COMMIT: string;
+}
