+++ web-local/src/lib/components/layouts/index.ts
@@ -1,0 +1,1 @@
+export { default as RillDeveloperApplicationLayout } from "./RillDeveloperApplicationLayout.svelte";
