+++ web-local/src/routes/(application)/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import { RillDeveloperApplicationLayout } from "@rilldata/web-local/lib/components/layouts";
+</script>
+
+<RillDeveloperApplicationLayout>
+  <slot />
+</RillDeveloperApplicationLayout>
