+++ web-local/src/routes/(application)/dashboard/[id]/+page.svelte
