+++ web-local/src/routes/(application)/model/[id]/+page.svelte
